Layout fix complete

+++ app/src/main/java/com/kelly/coinplace/data/remote/dto/CoinsDto.kt
@@ -4,16 +4,16 @@ import com.kelly.coinplace.domain.model.Coins
 
 data class CoinsDto(
     val beta_value: Double,
-    val circulating_supply: Int,
+    val circulating_supply: Long,
     val first_data_at: String,
     val id: String,
     val last_updated: String,
-    val max_supply: Int,
+    val max_supply: Long,
     val name: String,
     val quotes: Quotes,
     val rank: Int,
     val symbol: String,
-    val total_supply: Int
+    val total_supply: Long
 )
 
 data class Quotes(
